server/requirements.txt
fastapi
uvicorn
websockets
langchain
langserve
python-dotenv
langchain_openai
